Revert "Feature/button update button"
