--- src/Component/BlazorComponent/Components/Button/Components/Theme.cs
@@ -1,10 +1,0 @@
-﻿namespace BlazorComponent;
-
-public enum Theme
-{
-    Solid,
-
-    Borderless,
-
-    Light
-}
