+++ src/Component/BlazorComponent/Components/Button/SButton.razor.cs
@@ -58,7 +58,7 @@ public partial class SButton
 
         if (Disabled)
         {
-            CssProvider?.CssApply(PrefixCls+"-disabled");
+            CssProvider?.CssApply("semi-disabled");
         }
 
         CssProvider.CssApply(PrefixCls);
@@ -70,12 +70,12 @@ public partial class SButton
 
         if (!string.IsNullOrEmpty(Size))
         {
-            CssProvider?.CssApply(PrefixCls + "-size-" + Size.ToLower());
+            CssProvider?.CssApply(PrefixCls + Size.ToLower());
         }
 
         if (!string.IsNullOrEmpty(Theme))
         {
-            CssProvider?.CssApply(PrefixCls + "-" + Theme.ToLower());
+            CssProvider?.CssApply(PrefixCls + Theme.ToLower());
         }
         else
         {
@@ -102,12 +102,6 @@ public partial class SButton
         {
             CssProvider?.CssApply(PrefixCls + "-primary");
         }
-
-        if (Block)
-        {
-            CssProvider?.CssApply(PrefixCls + "-block");
-        }
-
         base.OnInitialized();
     }
