--- src/Component/BlazorComponent/Components/Button/SIconButton.razor
@@ -1,62 +1,0 @@
-﻿@namespace BlazorComponent
-
-@inherits SDomComponentBase
-
-<SButton 
-    Secondary="@Secondary" 
-    Block="@Block" 
-    Circle="@Circle" 
-    Class="@CssProvider.GetClass()" 
-    Danger="@Danger"
-    Tertiary="@Tertiary" 
-    Theme="@Theme"
-    Attributes="@Attributes"
-    Size="@Size"
-    Style="@CssProvider.GetStyle()"
-    Warning="@Warning"
-    Id="@Id"
-    Disabled="@Disabled"
-    OnMouseDown="@OnMouseDown"
-    OnClick="@OnClick"
-    Light="@Light"
-    Link="@Link"
-    Height="@Height"
-    Key="@Key"
-    >
-    @if(IconPosition == "left")
-    {
-        @if (Icon != null || Loading && !Disabled)
-        {
-            // TODO: 需要加载器实现，
-            <SIconLoading></SIconLoading>
-        }
-        else
-        {
-            @Icon
-        }
-        if (@ChildContent != null)
-        {
-            <span class="@BtnTextCls.GetClass()" x-semi-prop="@XSemiProp">
-                @ChildContent
-            </span>
-        }
-    }
-    else
-    {
-        if (@ChildContent != null)
-        {
-            <span class="@BtnTextCls.GetClass()" x-semi-prop="@XSemiProp">
-                @ChildContent
-            </span>
-        }
-        @if (Icon != null || Loading && !Disabled)
-        {
-            // TODO: 需要加载器实现，
-            <SIconLoading></SIconLoading>
-        }
-        else
-        {
-            @Icon
-        }
-    }
-</SButton>
