--- src/Docs/Semi.Design.Shared/Component/Button/ButtonBlockLevel.razor
@@ -1,4 +1,0 @@
-﻿
-<div>
-    <SButton Block>块级按钮</SButton>
-</div>
