--- src/Docs/Semi.Design.Shared/Component/Button/ButtonControlStateDisabled.razor
@@ -1,8 +1,0 @@
-﻿
-<div>
-    <SButton Disabled>禁用</SButton>
-    <SButton Disabled Theme="borderless">无背景禁用</SButton>
-    <SButton Disabled Theme="light">浅色禁用</SButton>
-    <SButton Disabled Theme="borderless">无背景主要禁用</SButton>
-    <SButton Disabled Theme="solid" Warning>深色警告禁用</SButton>
-</div>
