--- src/Docs/Semi.Design.Shared/Component/Button/ButtonDeepBackground.razor
@@ -1,7 +1,0 @@
-﻿<div>
-    <SButton Theme='solid'  Style="margin-right: 8px">深色主要</SButton>
-    <SButton Theme='solid' Secondary Style="margin-right: 8px">深色次要</SButton>
-    <SButton Theme='solid' Tertiary Style="margin-right: 8px">深色第三</SButton>
-    <SButton Theme='solid' Warning Style="margin-right: 8px">深色警告</SButton>
-    <SButton Theme='solid' Danger Style="margin-right: 8px">深色危险</SButton>
-</div>
