--- src/Docs/Semi.Design.Shared/Component/Button/ButtonGroupType.razor
@@ -1,42 +1,0 @@
-﻿
-
-<div style="display: flex">
-
-    <div style="margin-right: 10px" key="primary">
-        <SButtonGroup>
-            <SButton>拷贝</SButton>
-            <SButton>查询</SButton>
-            <SButton>剪切</SButton>
-        </SButtonGroup>
-    </div>
-
-    <div style="margin-right: 10px" key="Secondary">
-        <SButtonGroup>
-            <SButton Secondary>拷贝</SButton>
-            <SButton Secondary>查询</SButton>
-            <SButton Secondary>剪切</SButton>
-        </SButtonGroup>
-    </div>
-    <div style="margin-right: 10px" key="Tertiary">
-        <SButtonGroup>
-            <SButton Tertiary>拷贝</SButton>
-            <SButton Tertiary>查询</SButton>
-            <SButton Tertiary>剪切</SButton>
-        </SButtonGroup>
-    </div>
-    <div style="margin-right: 10px" key="warning">
-        <SButtonGroup>
-            <SButton Warning>拷贝</SButton>
-            <SButton Warning>查询</SButton>
-            <SButton Warning>剪切</SButton>
-        </SButtonGroup>
-    </div>
-    <div style="margin-right: 10px" key="Danger">
-        <SButtonGroup>
-            <SButton Danger>拷贝</SButton>
-            <SButton Danger>查询</SButton>
-            <SButton Danger>剪切</SButton>
-        </SButtonGroup>
-    </div>
-
-</div>
