--- src/Docs/Semi.Design.Shared/Component/Button/ButtonIcon.razor
@@ -1,65 +1,0 @@
-﻿
-<div>
-    <strong>默认状态：</strong>
-    <SIconButton aria-label="截屏">
-        <Icon>
-            <SIconCamera />
-        </Icon>
-    </SIconButton>
-    <br /><br />
-    <strong>禁用状态：</strong>
-    <SIconButton Disabled aria-label="截屏">
-        <Icon><SIconCamera /></Icon>
-    </SIconButton>
-    <br /><br />
-    <strong>复合类型：</strong>
-    <span className="btn-margin-right">
-
-        <SIconButton Disabled aria-label="截屏">
-            <Icon><SIconCamera /></Icon>
-        </SIconButton>
-
-
-        <SIconButton Disabled aria-label="截屏">
-            <Icon><SIconCamera /></Icon>
-        </SIconButton>
-
-
-        <SIconButton Disabled aria-label="截屏">
-            <Icon><SIconCamera /></Icon>
-        </SIconButton>
-
-        <SIconButton Warning Disabled aria-label="截屏">
-            <Icon><SIconCamera /></Icon>
-        </SIconButton>
-
-        <SIconButton Danger Disabled aria-label="截屏">
-            <Icon><SIconCamera /></Icon>
-        </SIconButton>
-
-    </span>
-    <br /><br />
-    <strong>更改主题：</strong>
-
-    <SIconButton Theme="solid" Style="margin-right:10px" aria-label="截屏">
-        <Icon><SIconCamera /></Icon>
-    </SIconButton>
-
-    <SIconButton Theme="light" Style="margin-right:10px" aria-label="截屏">
-        <Icon><SIconCamera /></Icon>
-    </SIconButton>
-
-    <br /><br />
-    <strong>更改图标位置：</strong>
-
-    <SIconButton Theme="solid" Style="margin-right:10px" aria-label="截屏">
-        <Icon><SIconSidebar /></Icon>
-        <ChildContent>收起</ChildContent>
-    </SIconButton>
-    <SIconButton Theme="solid" IconPosition="right" Style="margin-right:10px" aria-label="截屏">
-        <Icon><SIconChevronDown /></Icon>
-        <ChildContent>展开选项</ChildContent>
-    </SIconButton>
-
-    <br /><br />
-</div>
