--- src/Docs/Semi.Design.Shared/Component/Button/ButtonLoadState.razor
@@ -1,54 +1,0 @@
-﻿
-<div>
-    <div>
-        <div className="btn-margin-right" style="display: inline-flex; align-items: center; padding-bottom: 14px;">
-            <SButton OnClick="() => ReSet(false)">关闭加载态</SButton>
-            <SButton OnClick="() => ReSet(true)">开启加载态</SButton>
-        </div>
-    </div>
-    <hr />
-    <SIconButton Loading="SaveLoading" OnClick="()=> SetSaveLoading(true)" Style="margin-right:14px">保存</SIconButton>
-
-    <SIconButton Loading="DelLoading" OnClick="() => SetDelLoading(true)" Danger Style="margin-right: 14px">
-        <Icon>
-            <SIconDelete/>
-        </Icon>
-        <ChildContent>
-            删除
-        </ChildContent>
-    </SIconButton>
-    
-    <div style="width: 200px;display:inline-block">
-        <SIconButton Loading="RepLoading" Warning Block Theme="solid" OnClick="()=> SetRepLoading(true)">撤销</SIconButton>
-    </div>
-</div>
-
-@code
-{
-    private bool Status;
-    private bool SaveLoading;
-    private bool DelLoading;
-    private bool RepLoading;
-
-    private void SetSaveLoading(bool saveLoading)
-    {
-        SaveLoading = saveLoading;
-        StateHasChanged();
-    }
-    private void SetDelLoading(bool delLoading)
-    {
-        DelLoading = delLoading;
-        StateHasChanged();
-    }
-    private void SetRepLoading(bool repLoading)
-    {
-        RepLoading = repLoading;
-        StateHasChanged();
-    }
-
-    private void ReSet(bool status)
-    {
-        Status = status;
-        StateHasChanged();
-    }
-}
