--- src/Docs/Semi.Design.Shared/Component/Button/ButtonSize.razor
@@ -1,6 +1,0 @@
-﻿
-<div>
-    <SButton Size='large' Style="margin-right: 8px">大尺寸</SButton>
-    <SButton Size='default' Style="margin-right: 8px">默认尺寸</SButton>
-    <SButton Size='small'>小尺寸</SButton>
-</div>
