--- src/Docs/Semi.Design.Shared/Component/Button/ButtonType.razor
@@ -1,8 +1,0 @@
-﻿
-<div className="btn-margin-right">
-    <SButton>主要按钮</SButton>
-    <SButton Secondary>次要按钮</SButton>
-    <SButton Tertiary>第三按钮</SButton>
-    <SButton Warning>警告按钮</SButton>
-    <SButton Danger>危险按钮</SButton>
-</div>
