+++ src/Docs/Semi.Design.Shared/Component/MonacoEditor/SMonacoEditor.razor
@@ -5,12 +5,13 @@
     @ChildContent
 </div>
 
-<div class="@CssProvider.GetClass()" style="@CssProvider.GetStyle();" id="@Id" @ref="Ref" @attributes="@Attributes">
+<div class="@CssProvider.GetClass()" style="@CssProvider.GetStyle();float:left;" id="@Id" @ref="Ref" @attributes="@Attributes">
 </div>
 
 
 <style>
     .child{
         padding:8px;
+        float:left;
     }
 </style>
