+++ src/Docs/Semi.Design.Shared/Shared/MainLayout.razor
@@ -5,7 +5,7 @@
 <Menu MenuTree="@Navigation?.Menu" />
 
 <div class="content-area " id="main-content">
-    <div class="article-wrapper content">
+    <div class="content">
         @Body
     </div>
 </div>
