+++ src/Docs/Semi.Design.Shared/Shared/MainLayout.razor.css
@@ -1,4 +1,6 @@
 .content {
+    padding-left: 75px;
+    padding-right: 75px;
     padding-top:8px;
     padding-bottom:20px;
 }
