+++ src/Docs/Semi.Design.Shared/wwwroot/pages/Button/BasicUse.txt
@@ -1,0 +1,9 @@
+﻿
+    <h3>Index</h3>
+    <SButton>点击测试</SButton>
+
+    <SButtonGroup>
+        <SButton>测试</SButton>
+        <SButton>测试</SButton>
+        <SButton>测试</SButton>
+    </SButtonGroup>
